+++ fluttercraft/commands/__init__.py
@@ -7,6 +7,7 @@ from fluttercraft.commands.fvm_commands import (
     fvm_install_command,
     fvm_uninstall_command,
     fvm_releases_command,
+    fvm_list_command,
 )
 from fluttercraft.commands.help_commands import (
     show_global_help,
@@ -14,6 +15,7 @@ from fluttercraft.commands.help_commands import (
     show_fvm_install_help,
     show_fvm_uninstall_help,
     show_fvm_releases_help,
+    show_fvm_list_help,
     show_clear_help,
     handle_help_command,
 )